+++ crates/client/block_import/src/lib.rs
@@ -49,9 +49,9 @@ use std::{borrow::Cow, sync::Arc};
 mod metrics;
 mod pre_validate;
 mod rayon;
+pub mod tests;
 mod types;
 mod verify_apply;
-
 pub use pre_validate::*;
 pub use rayon::*;
 pub use types::*;