+++ crates/client/db/src/tests/test_open.rs
@@ -1,7 +1,5 @@
-mod common;
-
-use common::*;
-use mc_db::DatabaseService;
+use super::common::*;
+use crate::DatabaseService;
 use mc_metrics::MetricsRegistry;
 use mp_chain_config::ChainConfig;
 use mp_utils::tests_common::*;
@@ -10,7 +8,7 @@ use rstest::*;
 #[rstest]
 #[tokio::test]
 async fn test_open_db(_set_workdir: ()) {
-    temp_db().await;
+    temp_db::temp_db().await;
 }
 
 #[rstest]